allow anonymous user to user todo

+++ todo/views.py
@@ -2,11 +2,11 @@ from django.shortcuts import render,redirect
 from django.views.decorators.http import require_POST
 from .models import Todo
 from .forms import TodoForm
+from django.contrib.auth.decorators import login_required
 
 def index(request):
     
-    todo_list=Todo.objects.order_by('id')
-
+    todo_list=Todo.objects.order_by('id').filter(user_id=request.user.id)
     form =TodoForm()
 
     context={
@@ -15,11 +15,12 @@ def index(request):
     }
     return render(request,'todo/index.html',context)
 
+
 @require_POST
 def addTodo(request):
     form = TodoForm(request.POST)
     if form.is_valid():
-        new_todo=Todo(text=request.POST['text'])
+        new_todo=Todo(text=request.POST['text'],user_id=request.user.id)
         new_todo.save()
         return redirect('todo:index')
 
@@ -30,11 +31,11 @@ def completedTodo(request,todo_id):
     return redirect('todo:index')
 
 def deleteCompletedTodo(request):
-    Todo.objects.all().filter(completed=True).delete()
+    Todo.objects.all().filter(completed=True,user_id=request.user.id).delete()
     return redirect('todo:index')
 
 def deleteAll(request):
-    Todo.objects.all().delete()
+    Todo.objects.all().filter(user_id=request.user.id).delete()
     return redirect('todo:index')
